# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Diom API",
    "description": "Welcome to the Diom API documentation!\n",
    "version": "0.1.0",
    "x-logo": {
      "altText": "Svix Logo",
      "url": "https://www.svix.com/static/img/brand-padded.svg"
    }
  },
  "paths": {
...
    "/api/v1.admin.auth-role.configure": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Role Configure",
        "description": "Create or update a role",
        "operationId": "v1.admin.auth-role.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRoleConfigureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleConfigureOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.admin.auth-policy.configure": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Access Policy Configure",
        "description": "Create or update an access policy",
        "operationId": "v1.admin.auth-policy.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAccessPolicyConfigureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccessPolicyConfigureOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.auth-token.namespace.configure": {
      "post": {
        "tags": [
          "Auth Tokens"
        ],
        "summary": "Auth Token Configure Namespace",
        "description": "Configure Auth Token namespace",
        "operationId": "v1.auth-token.namespace.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthTokenConfigureNamespaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenConfigureNamespaceOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-internal": true
      }
    },
...
    "/api/v1.cache.namespace.configure": {
      "post": {
        "tags": [
          "Cache"
        ],
        "summary": "Cache Configure Namespace",
        "description": "Configure cache namespace",
        "operationId": "v1.cache.namespace.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheConfigureNamespaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheConfigureNamespaceOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.kv.namespace.configure": {
      "post": {
        "tags": [
          "Key Value Store"
        ],
        "summary": "Kv Configure Namespace",
        "description": "Configure KV namespace",
        "operationId": "v1.kv.namespace.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KvConfigureNamespaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KvConfigureNamespaceOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.rate-limit.namespace.configure": {
      "post": {
        "tags": [
          "Rate Limiter"
        ],
        "summary": "Rate Limit Configure Namespace",
        "description": "Configure rate limiter namespace",
        "operationId": "v1.rate-limit.namespace.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateLimitConfigureNamespaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitConfigureNamespaceOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.idempotency.namespace.configure": {
      "post": {
        "tags": [
          "Idempotency"
        ],
        "summary": "Idempotency Configure Namespace",
        "description": "Configure idempotency namespace",
        "operationId": "v1.idempotency.namespace.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdempotencyConfigureNamespaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdempotencyConfigureNamespaceOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.msgs.namespace.configure": {
      "post": {
        "tags": [
          "Msgs"
        ],
        "summary": "Configure Namespace",
        "description": "Configures a msgs namespace with the given name.",
        "operationId": "v1.msgs.namespace.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MsgNamespaceConfigureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MsgNamespaceConfigureOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.msgs.queue.configure": {
      "post": {
        "tags": [
          "Msgs"
        ],
        "summary": "Queue Configure",
        "description": "Configures retry and DLQ behavior for a consumer group on a topic.\n\n`retry_schedule` is a list of delays (in millis) between retries after a nack. Once exhausted,\nthe message is moved to the DLQ (or forwarded to `dlq_topic` if set).",
        "operationId": "v1.msgs.queue.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MsgQueueConfigureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MsgQueueConfigureOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1.msgs.topic.configure": {
      "post": {
        "tags": [
          "Msgs"
        ],
        "summary": "Topic Configure",
        "description": "Configures the number of partitions for a topic.\n\nPartition count can only be increased, never decreased. The default for a new topic is 1.",
        "operationId": "v1.msgs.topic.configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MsgTopicConfigureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MsgTopicConfigureOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "HTTP Bearer token passed in the `Authorization` header"
      }
    },
    "schemas": {
...
      "AdminAccessPolicyConfigureIn": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccessRule"
            },
            "default": []
          }
        },
        "required": [
          "id",
          "description"
        ]
      },
      "AdminAccessPolicyConfigureOut": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "created": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          },
          "updated": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          }
        },
        "required": [
          "id",
          "created",
          "updated"
        ]
      },
...
      "AdminRoleConfigureIn": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccessRule"
            },
            "default": []
          },
          "policies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "context": {
            "type": "object",
            "default": {},
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "description"
        ]
      },
      "AdminRoleConfigureOut": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "created": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          },
          "updated": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          }
        },
        "required": [
          "id",
          "created",
          "updated"
        ]
      },
...
      "AuthTokenConfigureNamespaceIn": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "pattern": "^[a-zA-Z0-9\\-/_.=+]+$",
            "example": "some_namespace"
          }
        },
        "required": [
          "name"
        ]
      },
      "AuthTokenConfigureNamespaceOut": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "pattern": "^[a-zA-Z0-9\\-/_.=+]+$",
            "example": "some_namespace"
          },
          "created": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          },
          "updated": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          }
        },
        "required": [
          "name",
          "created",
          "updated"
        ]
      },
...
      "CacheConfigureNamespaceIn": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "pattern": "^[a-zA-Z0-9\\-/_.=+]+$",
            "example": "some_namespace"
          },
          "eviction_policy": {
            "default": "no-eviction",
            "$ref": "#/components/schemas/EvictionPolicy"
          }
        },
        "required": [
          "name"
        ]
      },
      "CacheConfigureNamespaceOut": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "pattern": "^[a-zA-Z0-9\\-/_.=+]+$",
            "example": "some_namespace"
          },
          "eviction_policy": {
            "$ref": "#/components/schemas/EvictionPolicy"
          },
          "created": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          },
          "updated": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0,
            "x-subtype": "UnixTimestampMs"
          }
        },
        "required": [
          "name",
          "eviction_policy",
          "created",
          "updated"
        ]
      },
...
      "EvictionPolicy": {
        "type": "string",
        "enum": [
          "no-eviction"
        ]
      },
...
      "StandardErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ]
      },
...
      "ValidationErrorBody": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        },
        "required": [
          "detail"
        ]
      },
...
    }
  },
...
}